chore: 🤖 Restructure the components directory
Modifies the structure of the components directory, placing interface
sections at the first level of components
